requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
contextlib2==21.6.0
crontab==0.23.0
Deprecated==1.2.13
Django==4.1
django-async-orm==0.1.10
django-environ==0.9.0
django-ninja==0.19.1
django-ninja-extra==0.15.4
djangorestframework==3.13.1
greenlet==1.1.2
idna==3.3
injector==0.19.0
kombu==5.2.4
mysql-connector-python==8.0.30
mysqlclient==2.1.1
packaging==21.3
Pillow==9.2.0
prompt-toolkit==3.0.30
protobuf==3.20.1
pydantic==1.9.1
PyMySQL==1.0.2
pyparsing==3.0.9
python-dotenv==0.20.0
pytz==2022.1
redis==4.3.4
requests==2.28.1
six==1.16.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
typing_extensions==4.3.0
tzdata==2022.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
